# Backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Översvämningskollen API
  version: 1.0.0
  description: API för hantering av användare, sensorer och vattennivåer.
servers:
  - url: http://localhost:3000
paths:
...
  /api/emergency-contacts:
    get:
      summary: Hämta alla nödkontakter
      tags:
        - Emergency Contacts
      responses:
        '200':
          description: En lista med alla kontakter
        '500':
          description: Fel vid hämtning av kontakter

    post:
      summary: Skapa en ny nödkontakt
      tags:
        - Emergency Contacts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sensor_id
                - name
                - phone_number
              properties:
                sensor_id:
                  type: integer
                name:
                  type: string
                phone_number:
                  type: string
      responses:
        '201':
          description: Kontakt tillagd
        '400':
          description: Ogiltig inmatning
        '500':
          description: Fel vid skapande av kontakt
...